Annotation and Style Improvements: Part I (#14)
* Annotation and style improvements wip

* Adding annotations in progress

* Annotate ddpg

* Annotate SAC and TQC

* refactor: state input dimenstion changed to the single int

* chore: update mujoco and pytorch versions

* fix: palce gaussian actor for sac and tqc on device properly

+++ src/oprl/env.py
@@ -96,7 +96,7 @@ class DMControlEnv:
                 obs_flatten.append(o.flatten())
             else:
                 obs_flatten.append(o)
-        return np.concatenate(obs_flatten)
+        return np.concatenate(obs_flatten, dtype="float32")
 
     @property
     def env_family(self) -> str:
